\name{namespace-package}
\alias{namespace-package}
\alias{namespace}
\docType{package}
\title{Namespace managment functions}
\description{
This package provides user-level functions to manage namespaces not
(yet) available in base R.
}
\details{
\tabular{ll}{
Package: \tab namespace\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-13\cr
License: \tab GPL-2\cr
}
}
\author{
Winston Chang, Daniel Adler, Hadley Wickham, Gregory R. Warnes, some
code from the R 'base' package.

Maintainer: Gregory R. Warnes <greg@warnes.net>
}
\keyword{ package }
\keyword{ programming }
\seealso{
  \code{\link[base]{ns-internal}}
}
